Rename cpp files
